# Optimize `gen(::MPolyRing, ::Int)` for :deglex and :degrevlex