Before:

    julia> R, _ = polynomial_ring(QQ, 2);

    julia> @b gen($R, 1)
    79.036 ns (9 allocs: 304 bytes)

    julia> R, _ = polynomial_ring(QQ, 10; internal_ordering=:degrevlex);

    julia> @b gen($R, 1)
    613.239 ns (21 allocs: 752 bytes)

After:

    julia> R, _ = polynomial_ring(QQ, 2);

    julia> @b gen($R, 1)
    71.306 ns (9 allocs: 304 bytes)

    julia> R, _ = polynomial_ring(QQ, 10; internal_ordering=:degrevlex);

    julia> @b gen($R, 1)
    75.127 ns (9 allocs: 368 bytes)